Move block tree testing utilities from pallet_domains::block_tree::tests to pallet_domains::tests
These utilities will also be used by the upcoming fraud proof test, so moving
them to an upper module to share
